requirements.txt
# Core ML/DL Framework
torch==2.1.0
torchvision==0.16.0 # Ensure compatibility with torch 2.1.0

# Monitoring & Profiling
pynvml>=11.5.0     # For NVIDIA GPU monitoring
torchprof>=1.0.0   # For FLOPs profiling (adjust version as needed)

# Experiment Tracking & Hyperparameter Optimization
wandb>=0.15.0      # Weights & Biases
optuna>=3.0.0      # Hyperparameter optimization

# Configuration & Utilities
PyYAML>=6.0        # For reading YAML configs
numpy>=1.21.0      # Fundamental package for scientific computing
scikit-learn>=1.0  # For potential utility functions like train/val split if needed beyond simple slicing

# Add other specific dependencies as they arise
